+++ config.py
@@ -1,9 +1,9 @@
 class DefaultConfig(object):
 
     data_params = {
-        "dataset_root": r'augmented_data_root',
-        "pretrain_data_root": r'augmented_data_root',
-        "finetune_data_root": r'clinical_data_root',
+        "dataset_root": r'dataset/train',
+        "pretrain_data_root": r'dataset/train',
+        "finetune_data_root": r'dataset/train',
         "eos_coef": 0.2,
         "train_ratio": 0.7,  # 70/15/15 split (train/validation/test)
         "window_lw": [300, 900],